Commit Aaron

+++ ADM.Models/AbonoCXP.cs
@@ -6,9 +6,9 @@
         public string FK_Documento { get; set; }
         public string FK_Proveedor { get; set; }
         public DateTime Fecha_Documento { get; set; }
-        public decimal Monto { get; set; }
-        public decimal Saldo_Pendiente { get; set; }
-        public decimal Monto_Abonado { get; set; }
+        public double Monto { get; set; }
+        public double Saldo_Pendiente { get; set; }
+        public double Monto_Abonado { get; set; }
         public string Tipo_Documento { get; set; }
         public string Banco { get; set; }
         public string? FK_Usuario_Creacion { get; set; }
